add DoxygenLayout.xml to doxygen script

+++ Src/memory/src/tables.cpp
@@ -16,7 +16,7 @@ int32_t tables::get_value(table_ref table, uint32_t x, uint32_t y) {
   return (int32_t)(data[1] << 8) + (data[2] << 16) + (data[3] << 24) + data[0];
 }
 
-void tables::set_value(table_ref table, uint16_t x, uint16_t y, int32_t value) {
+void tables::set_value(table_ref table, uint32_t x, uint32_t y, int32_t value) {
   uint32_t address = get_address(x, table.x_max, y, table.memory_address);
 
   uint8_t data[4];
